backend/requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
Flask-JWT-Extended>=4.4
Flask-CORS>=4.0
passlib[bcrypt]>=1.7
python-dotenv>=1.0
SQLAlchemy>=1.4
psycopg2-binary>=2.9
alembic>=1.8
gunicorn>=20.1
redis>=4.5
flasgger>=0.9.7

# ML Dependencies
scikit-learn>=1.3.0
joblib>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
lightgbm>=4.0.0

# Background Tasks
celery>=5.3.0
celery[redis]>=5.3.0

# Monitoring & Utilities
python-dateutil>=2.8.0
